refactor(core.js): remove /core/toolkit.js
